Corrected mis-placed else for enter IP section - for when a setting is not set etc.
Changed Client to local variable for now, prevents an error caused by public variable and disconnections.
Added a brief sleep time to the code to allow for the tiny delay server side and actually getting the data sent back.
Remove the starting phrase on the protocol (????print)
Added while loop to cope with multiple packets  (currently fails to work properly not sure why, can't test without a full server.
Set the buffer size to precisely what we are going to be grabbing from the connection.